Debug des swipes et ajout des chats

+++ app/src/main/java/com/e_swipe/e_swipe/model/UserCard.java
@@ -49,10 +49,10 @@ public class UserCard {
     }
 
     public class position {
-        int latitude;
-        int longitude;
+        float latitude;
+        float longitude;
 
-        position(int latitude,int longitude){
+        position(float latitude,float longitude){
             this.latitude = latitude;
             this.longitude = longitude;
         }
